snowflake

+++ src/main/java/com/iboot/admin/repository/GroupRepository.java
@@ -5,7 +5,7 @@ import com.iboot.core.repository.BaseRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupRepository extends BaseRepository<Group,Integer> {
+public interface GroupRepository extends BaseRepository<Group,Long> {
 
 
 }
